fix: separate line and normal emoticon commands
